delivery/requirements.txt
altgraph==0.17.2
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
Flask==2.2.2
future==0.18.2
idna==3.4
importlib-metadata==5.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
macholib==1.15.2
MarkupSafe==2.1.1
pip==21.2.4
setuptools==58.0.4
six==1.15.0
urllib3==1.26.13
Werkzeug==2.2.2
wheel==0.37.0
zipp==3.11.0
uuid==1.30
redis==4.3.5
requests==2.28.1
opentelemetry-exporter-otlp-proto-http
opentelemetry-instrumentation
opentelemetry-instrumentation-flask
opentelemetry-instrumentation-requests
opentelemetry.instrumentation.logging
